+++ PantherGUI/container.h
@@ -16,7 +16,7 @@ public:
 	PGContainer(PGWindowHandle window);
 	virtual ~PGContainer();
 
-	virtual void MouseWheel(int x, int y, int distance, PGModifier modifier);
+	virtual void MouseWheel(int x, int y, double distance, PGModifier modifier);
 	virtual bool KeyboardButton(PGButton button, PGModifier modifier);
 	virtual bool KeyboardCharacter(char character, PGModifier modifier);
 	virtual bool KeyboardUnicode(PGUTF8Character character, PGModifier modifier);
